Refactor massive entry file into smaller modules
