+++ bamboo-core/src/entry/mod.rs
@@ -1,36 +1,31 @@
-use crate::BLAKE2B_HASH_SIZE;
 use arrayvec::ArrayVec;
 use core::borrow::Borrow;
 use core::convert::TryFrom;
-use ed25519_dalek::PublicKey;
 use lipmaa_link::lipmaa;
-use std::collections::HashMap;
+
+pub mod decode;
+pub mod encode;
+pub mod publish;
+pub mod verify;
+pub mod verify_batch;
+
+pub use decode::decode;
+pub use verify::verify;
+pub use verify_batch::verify_batch;
+pub use publish::publish;
 
 #[cfg(feature = "std")]
 use crate::util::hex_serde::*;
 
 #[cfg(feature = "std")]
-use std::io::Write;
-use varu64::{
-    decode as varu64_decode, encode as varu64_encode, encoding_length as varu64_encoding_length,
-};
-
-#[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
-#[cfg(feature = "std")]
-use varu64::encode_write as varu64_encode_write;
 
 use ed25519_dalek::{
-    PublicKey as DalekPublicKey, Signature as DalekSignature, Signer, Verifier, PUBLIC_KEY_LENGTH,
+    PublicKey as DalekPublicKey, PUBLIC_KEY_LENGTH,
 };
-
-#[cfg(feature = "std")]
-use ed25519_dalek::verify_batch;
 
 use super::signature::{Signature, MAX_SIGNATURE_SIZE};
 use super::yamf_hash::{YamfHash, MAX_YAMF_HASH_SIZE};
-use crate::yamf_hash::new_blake2b;
-use ed25519_dalek::Keypair;
 
 pub use crate::error::*;
 
@@ -105,456 +100,6 @@ where
     }
 }
 
-pub fn verify_links_and_payload(
-    entry: &Entry<&[u8], &[u8]>,
-    payload: Option<(&[u8], YamfHash<ArrayVec<[u8; 64]>>)>,
-    lipmaa_link: Option<(&[u8], YamfHash<ArrayVec<[u8; 64]>>)>,
-    backlink: Option<(&[u8], YamfHash<ArrayVec<[u8; 64]>>)>,
-) -> Result<(), Error> {
-    // If we have the payload, check that its hash and length match what is encoded in the
-    // entry.
-    if let Some((payload, payload_hash)) = payload {
-        if payload_hash != entry.payload_hash {
-            return Err(Error::AddEntryPayloadHashDidNotMatch);
-        }
-        if payload.len() as u64 != entry.payload_size {
-            return Err(Error::AddEntryPayloadLengthDidNotMatch);
-        }
-    }
-
-    let lipmaa_is_required = is_lipmaa_required(entry.seq_num);
-
-    match (
-        lipmaa_link,
-        entry.lipmaa_link.as_ref(),
-        entry.seq_num,
-        lipmaa_is_required,
-    ) {
-        // Happy path 1: this is the first entry, so we won't find a lipmaa link in the store
-        (None, None, seq_num, _) if seq_num == 1 => Ok(()),
-        // Happy path 2: seq is larger than one and we can find the lipmaa link in the store
-        (Some((lipmaa, lipmaa_hash)), Some(ref entry_lipmaa), seq_num, true) if seq_num > 1 => {
-            // Make sure the lipmaa entry hash matches what's in the entry.
-            if lipmaa_hash != **entry_lipmaa {
-                return Err(Error::AddEntryLipmaaHashDidNotMatch);
-            }
-
-            let lipmaa_entry =
-                decode(lipmaa).map_err(|_| Error::AddEntryDecodeLipmaalinkFromStore)?;
-
-            // Verify that the log_id of the entry is the same as the lipmaa entry
-            if entry.log_id != lipmaa_entry.log_id {
-                return Err(Error::AddEntryLogIdDidNotMatchLipmaaEntry);
-            }
-            // Verify the author of the entry is the same as the author in the lipmaa link entry
-            if entry.author != lipmaa_entry.author {
-                return Err(Error::AddEntryAuthorDidNotMatchLipmaaEntry);
-            }
-            Ok(())
-        }
-        // Happy path 3: lipmaa link is not required because it would duplicate the backlink.
-        (_, _, seq_num, false) if seq_num > 1 => Ok(()),
-        (_, _, _, _) => Err(Error::AddEntryNoLipmaalinkInStore),
-    }?;
-
-    match (backlink, entry.backlink.as_ref(), entry.seq_num) {
-        // Happy path 1: This is the first entry and doesn't have a backlink.
-        (_, None, seq_num) if seq_num == 1 => Ok(()),
-
-        //Happy path 2: This does have a backlink and we found it.
-        (Some((backlink, backlink_hash)), Some(ref entry_backlink), seq_num) if seq_num > 1 => {
-            let backlink_entry = decode(backlink).map_err(|_| Error::AddEntryDecodeLastEntry)?;
-
-            // Verify that the log_id of the entry is the same as the lipmaa entry
-            if entry.log_id != backlink_entry.log_id {
-                return Err(Error::AddEntryLogIdDidNotMatchPreviousEntry);
-            }
-            // Verify the author of the entry is the same as the author in the lipmaa link entry
-            if entry.author != backlink_entry.author {
-                return Err(Error::AddEntryAuthorDidNotMatchPreviousEntry);
-            }
-
-            // Verify this wasn't published after an end of feed message.
-            if backlink_entry.is_end_of_feed {
-                return Err(Error::AddEntryToFeedThatHasEnded);
-            }
-
-            if backlink_hash != **entry_backlink {
-                return Err(Error::AddEntryBacklinkHashDidNotMatch);
-            }
-            Ok(())
-        }
-        //Happy path 3: We don't have the backlink for this entry, happens when doing partial
-        //replication.
-        (None, Some(_), seq_num) if seq_num > 1 => Ok(()),
-        (_, _, _) => Err(Error::AddEntryBacklinkHashDidNotMatch),
-    }?;
-
-    Ok(())
-}
-
-pub fn verify(
-    entry_bytes: &[u8],
-    payload: Option<&[u8]>,
-    lipmaa_link: Option<&[u8]>,
-    backlink: Option<&[u8]>,
-) -> Result<bool, Error> {
-    // Decode the entry that we want to add.
-    let entry = decode(entry_bytes).map_err(|_| Error::AddEntryDecodeFailed)?;
-
-    let payload_and_hash = payload.map(|payload| (payload, new_blake2b(payload)));
-    let lipmaa_link_and_hash = lipmaa_link.map(|link| (link, new_blake2b(link)));
-    let backlink_and_hash = backlink.map(|link| (link, new_blake2b(link)));
-
-    verify_links_and_payload(
-        &entry,
-        payload_and_hash,
-        lipmaa_link_and_hash,
-        backlink_and_hash,
-    )?;
-
-    let is_valid = entry
-        .verify_signature()
-        .map_err(|_| Error::AddEntrySigNotValidError)?;
-
-    Ok(is_valid)
-}
-
-pub fn publish(
-    out: &mut [u8],
-    key_pair: Option<&Keypair>,
-    log_id: u64,
-    payload: &[u8],
-    is_end_of_feed: bool,
-    last_seq_num: Option<u64>,
-    lipmaa_entry_bytes: Option<&[u8]>,
-    backlink_bytes: Option<&[u8]>,
-) -> Result<usize, Error> {
-    let author = key_pair
-        .as_ref()
-        .map(|keys| keys.public.clone())
-        .ok_or(Error::PublishWithoutKeypair)?;
-
-    // calc the payload hash
-    let payload_hash = new_blake2b(payload);
-    let payload_size = payload.len() as u64;
-
-    let seq_num = last_seq_num.unwrap_or(0) + 1;
-
-    let mut entry: Entry<_, &[u8]> = Entry {
-        log_id,
-        is_end_of_feed,
-        payload_hash,
-        payload_size,
-        author,
-        seq_num,
-        backlink: None,
-        lipmaa_link: None,
-        sig: None,
-    };
-
-    // if the seq is larger than 1, we need to append the lipmaa and backlink hashes.
-    if seq_num > 1 {
-        let lipmaa_link = new_blake2b(lipmaa_entry_bytes.ok_or(Error::PublishWithoutLipmaaEntry)?);
-
-        //Make sure we're not trying to publish after the end of a feed.
-        let backlink_entry =
-            decode(&backlink_bytes.ok_or(Error::PublishWithoutBacklinkEntry)?[..])?;
-        if backlink_entry.is_end_of_feed {
-            return Err(Error::PublishAfterEndOfFeed);
-        }
-
-        // Avoid publishing to a feed using an incorrect log_id
-        if log_id != backlink_entry.log_id {
-            return Err(Error::PublishWithIncorrectLogId);
-        }
-
-        let backlink = new_blake2b(backlink_bytes.ok_or(Error::PublishWithoutBacklinkEntry)?);
-        entry.backlink = Some(backlink);
-
-        // If the lipmaalink and backlink would be different, we should append the lipmaalink,
-        // otherwise we're allowed to omit it to save some bytes.
-        if is_lipmaa_required(seq_num) {
-            entry.lipmaa_link = Some(lipmaa_link);
-        }
-    }
-
-    let mut buff = [0u8; 512];
-    let buff_size = entry.encode(&mut buff)?;
-
-    let signature = key_pair
-        .as_ref()
-        .ok_or(Error::PublishWithoutSecretKey)?
-        .sign(&buff[..buff_size]);
-    let sig_bytes = &signature.to_bytes()[..];
-    let signature = Signature(sig_bytes.into());
-
-    entry.sig = Some(signature);
-
-    entry.encode(out)
-}
-
-impl<'a, H, S> Entry<H, S>
-where
-    H: Borrow<[u8]>,
-    S: Borrow<[u8]>,
-{
-    pub fn encode(&self, out: &mut [u8]) -> Result<usize, Error> {
-        let mut next_byte_num = self.encode_for_signing(out)?;
-        // Encode the signature
-        if let Some(ref sig) = self.sig {
-            next_byte_num += sig
-                .encode(&mut out[next_byte_num..])
-                .map_err(|_| Error::EncodeSigError)?;
-        }
-
-        Ok(next_byte_num as usize)
-    }
-
-    pub fn encode_for_signing(&self, out: &mut [u8]) -> Result<usize, Error> {
-        if out.len() < self.encoding_length() {
-            return Err(Error::EncodeBufferLength);
-        }
-
-        let mut next_byte_num = 0;
-
-        // Encode the end of feed.
-        if self.is_end_of_feed {
-            out[0] = 1;
-        } else {
-            out[0] = 0;
-        }
-        next_byte_num += 1;
-
-        // Encode the author
-        let author_bytes = self.author.as_bytes();
-        out[next_byte_num..author_bytes.len() + next_byte_num].copy_from_slice(&author_bytes[..]);
-        next_byte_num += author_bytes.len();
-
-        // Encode the log_id
-        next_byte_num += varu64_encode(self.log_id, &mut out[next_byte_num..]);
-
-        // Encode the sequence number
-        next_byte_num += varu64_encode(self.seq_num, &mut out[next_byte_num..]);
-
-        // Encode the backlink and lipmaa links if its not the first sequence
-        next_byte_num = match (self.seq_num, &self.backlink, &self.lipmaa_link) {
-            (n, Some(ref backlink), Some(ref lipmaa_link)) if n > 1 => {
-                next_byte_num += lipmaa_link
-                    .encode(&mut out[next_byte_num..])
-                    .map_err(|_| Error::EncodeLipmaaError)?;
-                next_byte_num += backlink
-                    .encode(&mut out[next_byte_num..])
-                    .map_err(|_| Error::EncodeBacklinkError)?;
-                Ok(next_byte_num)
-            }
-            (n, Some(ref backlink), None) if n > 1 => {
-                next_byte_num += backlink
-                    .encode(&mut out[next_byte_num..])
-                    .map_err(|_| Error::EncodeBacklinkError)?;
-                Ok(next_byte_num)
-            }
-            (n, Some(_), Some(_)) if n <= 1 => Err(Error::EncodeEntryHasBacklinksWhenSeqZero),
-            _ => Ok(next_byte_num),
-        }?;
-
-        // Encode the payload size
-        next_byte_num += varu64_encode(self.payload_size, &mut out[next_byte_num..]);
-
-        // Encode the payload hash
-        next_byte_num += self
-            .payload_hash
-            .encode(&mut out[next_byte_num..])
-            .map_err(|_| Error::EncodePayloadHashError)?;
-
-        Ok(next_byte_num as usize)
-    }
-
-    /// Encode the entry ready for signing.
-    #[cfg(feature = "std")]
-    pub fn encode_for_signing_write<W: Write>(&self, mut w: W) -> Result<()> {
-        // Encode the "is end of feed" tag.
-        let mut is_end_of_feed_byte = [0];
-        if self.is_end_of_feed {
-            is_end_of_feed_byte[0] = 1;
-        }
-        w.write_all(&is_end_of_feed_byte[..])
-            .map_err(|_| Error::EncodeIsEndOfFeedError)?;
-
-        // Encode the author
-        let author_bytes = self.author.as_bytes();
-
-        w.write_all(author_bytes)
-            .map_err(|_| Error::EncodeAuthorError)?;
-
-        // Encode the log_id
-        varu64_encode_write(self.log_id, &mut w).map_err(|_| Error::EncodeLogIdError)?;
-
-        // Encode the sequence number
-        varu64_encode_write(self.seq_num, &mut w).map_err(|_| Error::EncodeSeqError)?;
-
-        // Encode the backlink and lipmaa links if its not the first sequence
-        match (self.seq_num, &self.backlink, &self.lipmaa_link) {
-            (n, Some(ref backlink), Some(ref lipmaa_link)) if n > 1 => {
-                lipmaa_link
-                    .encode_write(&mut w)
-                    .map_err(|_| Error::EncodeLipmaaError)?;
-
-                backlink
-                    .encode_write(&mut w)
-                    .map_err(|_| Error::EncodeBacklinkError)
-            }
-            (n, Some(ref backlink), None) if n > 1 => backlink
-                .encode_write(&mut w)
-                .map_err(|_| Error::EncodeBacklinkError),
-            (n, Some(_), Some(_)) if n <= 1 => Err(Error::EncodeEntryHasBacklinksWhenSeqZero),
-            _ => Ok(()),
-        }?;
-
-        // Encode the payload size
-        varu64_encode_write(self.payload_size, &mut w)
-            .map_err(|_| Error::EncodePayloadSizeError)?;
-
-        // Encode the payload hash
-        self.payload_hash
-            .encode_write(&mut w)
-            .map_err(|_| Error::EncodePayloadHashError)?;
-
-        Ok(())
-    }
-
-    #[cfg(feature = "std")]
-    pub fn encode_write<W: Write>(&self, mut w: W) -> Result<()> {
-        self.encode_for_signing_write(&mut w)?;
-
-        // Encode the signature
-        if let Some(ref sig) = self.sig {
-            sig.encode_write(&mut w)
-                .map_err(|_| Error::EncodeSigError)?;
-        }
-
-        Ok(())
-    }
-
-    pub fn encoding_length(&self) -> usize {
-        TAG_BYTE_LENGTH
-            + self.payload_hash.encoding_length()
-            + varu64_encoding_length(self.payload_size)
-            + varu64_encoding_length(self.log_id)
-            + self.author.as_bytes().len()
-            + varu64_encoding_length(self.seq_num)
-            + self
-                .backlink
-                .as_ref()
-                .map(|backlink| backlink.encoding_length())
-                .unwrap_or(0)
-            + self
-                .lipmaa_link
-                .as_ref()
-                .map(|lipmaa_link| lipmaa_link.encoding_length())
-                .unwrap_or(0)
-            + self
-                .sig
-                .as_ref()
-                .map(|sig| sig.encoding_length())
-                .unwrap_or(0)
-    }
-
-    pub fn verify_signature(&self) -> Result<bool> {
-        let ssb_sig = DalekSignature::try_from(self.sig.as_ref().unwrap().0.borrow())
-            .map_err(|_| Error::DecodeSsbSigError)?;
-
-        let mut buff = [0u8; 512];
-
-        let encoded_size = self.encode_for_signing(&mut buff).unwrap();
-
-        let pub_key = self.author.borrow();
-
-        let result = pub_key
-            .verify(&buff[..encoded_size], &ssb_sig)
-            .map(|_| true)
-            .unwrap_or(false);
-
-        Ok(result)
-    }
-}
-
-pub fn decode<'a>(bytes: &'a [u8]) -> Result<Entry<&'a [u8], &'a [u8]>, Error> {
-    // Decode is end of feed
-    if bytes.len() == 0 {
-        return Err(Error::DecodeInputIsLengthZero);
-    }
-    let is_end_of_feed = bytes[0] == 1;
-
-    // Decode the author
-    if bytes.len() < PUBLIC_KEY_LENGTH + 1 {
-        return Err(Error::DecodeAuthorError);
-    }
-
-    let author = DalekPublicKey::from_bytes(&bytes[1..PUBLIC_KEY_LENGTH + 1])
-        .map_err(|_| Error::DecodeAuthorError)?;
-
-    let remaining_bytes = &bytes[PUBLIC_KEY_LENGTH + 1..];
-
-    // Decode the log id
-    let (log_id, remaining_bytes) = varu64_decode(remaining_bytes)
-        .map_err(|(err, _)| err)
-        .map_err(|_| Error::DecodeLogIdError)?;
-
-    // Decode the sequence number
-    let (seq_num, remaining_bytes) = varu64_decode(remaining_bytes)
-        .map_err(|(err, _)| err)
-        .map_err(|_| Error::DecodeSeqError)?;
-
-    if seq_num == 0 {
-        return Err(Error::DecodeSeqIsZero);
-    }
-
-    let lipmaa_is_required = is_lipmaa_required(seq_num);
-
-    // Decode the backlink and lipmaa links if its not the first sequence
-    let (backlink, lipmaa_link, remaining_bytes) = match (seq_num, lipmaa_is_required) {
-        (1, _) => (None, None, remaining_bytes),
-        (_, true) => {
-            let (lipmaa_link, remaining_bytes) =
-                YamfHash::<&[u8]>::decode(remaining_bytes).map_err(|_| Error::DecodeLipmaaError)?;
-            let (backlink, remaining_bytes) = YamfHash::<&[u8]>::decode(remaining_bytes)
-                .map_err(|_| Error::DecodeBacklinkError)?;
-            (Some(backlink), Some(lipmaa_link), remaining_bytes)
-        }
-        (_, false) => {
-            let (backlink, remaining_bytes) = YamfHash::<&[u8]>::decode(remaining_bytes)
-                .map_err(|_| Error::DecodeBacklinkError)?;
-            (Some(backlink), None, remaining_bytes)
-        }
-    };
-
-    // Decode the payload size
-    let (payload_size, remaining_bytes) = varu64_decode(remaining_bytes)
-        .map_err(|(err, _)| err)
-        .map_err(|_| Error::DecodePayloadSizeError)?;
-
-    // Decode the payload hash
-    let (payload_hash, remaining_bytes) =
-        YamfHash::<&[u8]>::decode(remaining_bytes).map_err(|_| Error::DecodePayloadHashError)?;
-
-    // Decode the signature
-    let (sig, _) =
-        Signature::<&[u8]>::decode(remaining_bytes).map_err(|_| Error::DecodeSigError)?;
-
-    Ok(Entry {
-        log_id,
-        is_end_of_feed,
-        payload_hash,
-        payload_size,
-        author,
-        seq_num,
-        backlink,
-        lipmaa_link,
-        sig: Some(sig),
-    })
-}
-
 pub fn into_owned<H, S>(entry: &Entry<H, S>) -> Entry<ArrayVec<[u8; 64]>, ArrayVec<[u8; 64]>>
 where
     H: Borrow<[u8]>,
@@ -612,128 +157,4 @@ pub fn is_lipmaa_required(sequence_num: u64) -> bool {
     lipmaa(sequence_num) != sequence_num - 1
 }
 
-#[cfg(feature = "std")]
-use blake2b_simd::{
-    many::{hash_many, HashManyJob},
-    Params,
-};
 
-/// Batch verify a collection of entries that are **all from the same author and same log_id**
-#[cfg(feature = "std")]
-pub fn batch_verify<E: AsRef<[u8]>, P: AsRef<[u8]>>(entries_and_payloads: &[(E, Option<P>)]) -> Result<()> {
-    batch_verify_links_and_payload(entries_and_payloads)?;
-    let bytes_iter = entries_and_payloads.iter().map(|(bytes, _)| bytes.as_ref());
-    batch_verify_signatures(bytes_iter)?;
-
-    Ok(())
-}
-/// Batch verify the links + payloads of a collection of entries that are **all from the same author and same log_id**
-#[cfg(feature = "std")]
-pub fn batch_verify_links_and_payload<E: AsRef<[u8]>, P: AsRef<[u8]>>(
-    entries_and_payloads: &[(E, Option<P>)],
-) -> Result<()> {
-    let params = Params::new();
-
-    // Build a hashmap from seq num to bytes and hashes we need.
-    let mut hash_map = entries_and_payloads
-        .iter()
-        .map(|(bytes, payload)| {
-            let entry = Entry::try_from(bytes.as_ref())?;
-            let entry_job = HashManyJob::new(&params, bytes.as_ref());
-
-            let payload_and_job = payload.as_ref().map(|payload| {
-                (
-                    payload.as_ref(),
-                    HashManyJob::new(&params, payload.as_ref()),
-                )
-            });
-
-            Ok((
-                entry.seq_num,
-                (bytes.as_ref(), entry, entry_job, payload_and_job),
-            ))
-        })
-        .collect::<Result<HashMap<u64, (_, _, _, _)>>>()?;
-
-    // Hash all the entries at once.
-    hash_many(hash_map.iter_mut().map(|(_, (_, _, job, _))| job));
-
-    let payload_jobs = hash_map
-        .iter_mut()
-        .filter_map(|(_, (_, _, _, payload_and_job))| payload_and_job.as_mut().map(|(_, job)| job));
-
-    // Hash all the payloads at once.
-    hash_many(payload_jobs);
-
-    hash_map
-        .iter()
-        .map(|(seq_num, (_, entry, _, payload_and_job))| {
-            let backlink_and_hash = hash_map.get(&(seq_num - 1)).map(
-                |(bytes, _, entry_job, _)| -> (_, YamfHash<ArrayVec<[u8; BLAKE2B_HASH_SIZE]>>) {
-                    (*bytes, entry_job.to_hash().into())
-                },
-            );
-
-            let lipmaa_link_and_hash = hash_map.get(&(lipmaa_link::lipmaa(*seq_num))).map(
-                |(bytes, _, entry_job, _)| -> (_, YamfHash<ArrayVec<[u8; BLAKE2B_HASH_SIZE]>>) {
-                    (*bytes, entry_job.to_hash().into())
-                },
-            );
-
-            let payload_and_hash = payload_and_job
-                .as_ref()
-                .map(|(payload, job)| (*payload, job.to_hash().into()));
-
-            verify_links_and_payload(
-                entry,
-                payload_and_hash,
-                lipmaa_link_and_hash,
-                backlink_and_hash,
-            )
-        })
-        .collect()
-}
-
-/// Batch verify the signatures of a collection of entries that are **all from the same author and same log_id**
-#[cfg(feature = "std")]
-pub fn batch_verify_signatures<'a, I: IntoIterator<Item = &'a [u8]>>(
-    entries_bytes: I,
-) -> Result<()> {
-    let entries = entries_bytes
-        .into_iter()
-        .map(|bytes| Entry::try_from(bytes))
-        .collect::<Result<Vec<_>>>()?;
-
-    let unsigned_encoding_vecs = entries
-        .iter()
-        .map(|entry| {
-            let mut vec = Vec::with_capacity(entry.encoding_length());
-            entry.encode_for_signing_write(&mut vec)?;
-            Ok(vec)
-        })
-        .collect::<Result<Vec<Vec<u8>>>>()?;
-
-    let unsigned_encodings = unsigned_encoding_vecs
-        .iter()
-        .map(|entry| entry.as_ref())
-        .collect::<Vec<_>>();
-
-    let signatures = entries
-        .iter()
-        .map(|entry| {
-            let ssb_sig = DalekSignature::try_from(entry.sig.as_ref().unwrap().0.borrow())
-                .map_err(|_| Error::DecodeSsbSigError)?;
-            Ok(ssb_sig)
-        })
-        .collect::<Result<Vec<DalekSignature>>>()?;
-
-    let pub_keys = entries
-        .iter()
-        .map(|entry| entry.author.clone())
-        .collect::<Vec<PublicKey>>();
-
-    verify_batch(&unsigned_encodings, &signatures, &pub_keys[..])
-        .map_err(|_| Error::SignatureInvalid)?;
-
-    Ok(())
-}
